requirements.txt
llm
llm-ollama
ollama
insanely-fast-whisper
transformers
torch
accelerate
kokoro
scipy
python-obd